Remove new instrument query
