--- apps/frontend/src/graphql/instrument/getInstrumentsByRole.graphql
@@ -1,6 +1,0 @@
-query getInstrumentsByRole {
-  instrumentByRole {
-    ...instrument
-    
-  }
-}
